requirements.txt
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
anthropic>=0.18.0
rdflib>=6.0.0
streamlit>=1.35.0
pandas>=2.0.0
